@@ -18,7 +18,7 @@
     <name>cilantro</name>
     <description>Cilantro System Portal</description>
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <start-class>com.cilantro.CilantroApplication</start-class>
         <timestamp>${maven.build.timestamp}</timestamp>
         <maven.build.timestamp.format>yyyyMMdd</maven.build.timestamp.format>
